+++ test/runtests.jl
@@ -32,6 +32,7 @@ end
     @testset "resolver" include("test_resolver.jl")
     @testset "completions" include("test_completions.jl")
     @testset "signature help" include("test_signature_help.jl")
+    @testset "declaration" include("test_declaration.jl")
     @testset "definition" include("test_definition.jl")
     @testset "document highlight" include("test_document_highlight.jl")
     @testset "document symbol" include("test_document_symbol.jl")